% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmc_geo.R
\docType{data}
\name{idmc_geo}
\alias{idmc_geo}
\title{Geolocated internal displacements}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13450 rows and 6 columns.
}
\usage{
idmc_geo
}
\description{
Geolocated internal displacements
}
\keyword{datasets}
