% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco_occupations.R
\docType{data}
\name{isco_occupations}
\alias{isco_occupations}
\title{ISCO Occupations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 3 columns.
}
\source{
ISCO.
}
\usage{
isco_occupations
}
\description{
Occupational categories and corresponding skill level according to ISCO-08
and ISCO-68.
}
\details{
\describe{
\item{code}{ISCO code}
\item{label}{Occupational category}
\item{skill}{Skill level}
}

Occupational categories are from the International Standard
Classification of Occupations (ISCO), downloaded from
https://ilostat.ilo.org/methods/concepts-and-definitions/classification-occupation/.
}
\keyword{datasets}
