% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Grab data behind a preset plot}
\usage{
plot_data(key, iso = NULL, ...)
}
\arguments{
\item{key}{Plot key.}

\item{iso}{Country.}
}
\description{
Grab data behind a preset plot
}
