% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettylabel.R
\name{prettylabel}
\alias{prettylabel}
\title{Pretty number formatting}
\usage{
prettylabel(
  N,
  signif = 2,
  spell = FALSE,
  padzero = FALSE,
  pct = FALSE,
  currency = NULL
)
}
\arguments{
\item{N}{Number or vector of numbers to be formatted.}

\item{signif}{Number of significant digits.}

\item{spell}{Whether to spell out "millions", "billions", and "trillions".}

\item{padzero}{Whether to include zeros after the decimal point.}

\item{pct}{Whether to include a percent sign.}

\item{currency}{Whether to add a currency sign.}

\item{shorten}{Whether to abbreviate long numbers.}
}
\value{
String or vector of strings.
}
\description{
Formats numbers for display in charts.
}
