% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_digest.R
\name{render_digest}
\alias{render_digest}
\title{Render a Snapshot}
\usage{
render_digest(iso, target_dir = NULL, simplify_name = FALSE)
}
\arguments{
\item{iso}{Chosen country (ISO3).}

\item{target_dir}{Target directory. If \code{NULL}, saves output in the root
directory.}

\item{simplify_name}{Whether to simply file names.}
}
\value{
A PDF.
}
\description{
By default, output file name is the country's official name. If
\code{simplify_name} is set to \code{TRUE}, file name will be in all lowercase
without special characters and spaces will be replaced with "-".
}
