% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_snapshot}
\alias{repo_snapshot}
\alias{r_version}
\title{Download snapshot}
\usage{
repo_snapshot(
  repo,
  destdir = NULL,
  packages = NULL,
  win_binaries = FALSE,
  mac_binaries = FALSE,
  bin_versions = r_version()
)

r_version()
}
\arguments{
\item{repo}{url of the cran-like repository to snapshot}

\item{destdir}{directory in which to store the snapshot repository.}

\item{packages}{names of packages to include. Default is all packages in the repo.}

\item{win_binaries}{download binary packages for Windows}

\item{mac_binaries}{download binary packages for MacOS}

\item{bin_versions}{vector with versions of R to download the win/mac binary
packages. The default is to download binaries only for your local R version.}
}
\description{
Downloads a full copy (snapshot) of an R package repository. The resulting
folder can be hosted on any HTTP services to create a cran-like package
repository that can be used with `install.packages()` and others.
}
\examples{
repo_snapshot('https://jeroen.r-universe.dev', 'test', bin_versions = c("4.0", "4.1", "4.2", "4.3"))
unlink("test", recursive = TRUE)
}
