% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_plot.R
\name{test_plot}
\alias{test_plot}
\title{Test a plot template for all countries}
\usage{
test_plot(key, save_path = "test-plots", width = 10, height = 8, ...)
}
\arguments{
\item{key}{Plot key.}

\item{save_path}{Country.}

\item{...}{Optional arguments.}
}
\description{
Generate the \code{key} plot for all countries and save to
\code{save_path}.
}
